% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{ce_control}
\alias{ce_control}
\title{Control for Cross-Entropy Algorithm}
\usage{
ce_control(k, niter = 20, nsimul = 2000, sigma = diag(500, k),
  rho = 0.25)
}
\arguments{
\item{k}{Integer; dimension of price.}

\item{niter}{Integer; number of iterations.}

\item{nsimul}{Integer; number of simulations in one iteration.}

\item{sigma}{Numeric matrix (\code{k x k}, positive definite); initial variance matrix of prices.}

\item{rho}{Numeric (\code{0 < rho < 1}); how many (ratio of total) best performers to take.}
}
\value{
List with control parameters.
}
\description{
Control for Cross-Entropy Algorithm
}
\seealso{
\code{\link{find_next_price}}
}

