% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_all_routes.R
\name{drop_exceeding}
\alias{drop_exceeding}
\title{Drop exceeding passengers}
\usage{
drop_exceeding(tickets, occupancy, nseats)
}
\arguments{
\item{tickets}{Matrix containing information about the tickets.}

\item{occupancy}{Matrix of train occupancy after a passanger buys a ticket.
Rows are related to passengers, columns to pairs of consecutive stations.}

\item{nseat}{Number of seats in the train.}
}
\description{
Loops iteratively through all tickets and excludes passengers that do not fit the cappacity of a train.
}

