% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_intensity.R
\name{estimate_intensity}
\alias{estimate_intensity}
\title{Estimate intensity}
\usage{
estimate_intensity(data, ticket_price, routes, nseats)
}
\arguments{
\item{data}{}

\item{ticket_price}{}

\item{routes}{}

\item{nseats}{}
}
\description{
Estimates parameters of intensity function
}

