% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_next_price.R
\name{find_next_price}
\alias{find_next_price}
\title{Title}
\usage{
find_next_price(last_price, occupancy, time_breakpoint, routes, beta, nseats,
  init_new_price = 1.1 * last_price, ce = ce_control(k =
  length(last_price)), rs = rs_control())
}
\arguments{
\item{last_price}{Numeric vector.}

\item{occupancy}{Integer vector.}

\item{time_breakpoint}{Numeric vector.}

\item{routes}{Matrix.}

\item{beta}{Matrix.}

\item{nseats}{Integer.}

\item{init_new_price}{Numeric vector.}

\item{ce}{List. See \code{\link{ce_control}}.}

\item{rs}{List. See \code{\link{rs_control}}.}
}
\value{
Numeric vector of estimated optimal point.
}
\description{
Title
}
\examples{
last_price <- rep(50, 6)
occupancy <- rep(0, 3)
time_breakpoint <- 0:1
routes <- get_routes(4, 60)
beta <- matrix(rep(c(18,-4,4, 1.2), each = 6), nrow = 6)
nseats <- 1000

find_next_price(
  last_price, occupancy, time_breakpoint, routes, beta, nseats,
  init_new_price = 1.1 * last_price, ce, rs
)
}

