% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data}
\usage{
generate_data(beta, price, time_bp, ntrains, nstations, nseats, keep_all = F)
}
\arguments{
\item{beta}{Numeric matrix of model parameters. Each row contains parameters for one route.
The number of columns corresponds with the model selection. See details.}

\item{price}{Vector of ticket prices. Each element represents one route.}

\item{time_bp}{Vector of times when the price changes. Values between 0 and 1.}

\item{ntrains}{Number of datasets to be generated.}

\item{nstations}{Number of stations of the train.}

\item{nseat}{Number of seats in the train.}
}
\value{
List of datasets of simulated tickets.
Each element represents one train.
}
\description{
The function generates example data for the statistical model.
}
\examples{
beta <- matrix(rep(c(18,-4,4, 1.2), each = 6), nrow = 6)
price <- matrix(250, 2, 6)
generate_data <- function(beta, price, time_bp = 0.5, ntrains = 3, nstations = 4, nseats = 20)

}

