% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{rs_control}
\alias{rs_control}
\title{Control for Response Surface Algorithm}
\usage{
rs_control(niter = 200, nsimul = 500, max_step = 15, step_len = 1/20,
  sigma = 10, alpha = 0.3)
}
\arguments{
\item{niter}{Integer; number of iterations.}

\item{nsimul}{Integer; number of simulations in one iteration.}

\item{max_step}{Numeric; maximum change of price between two iterations in one dimension.}

\item{step_len}{Numeric; coefficient of step length. Estimated gradient is multiplied by this value.}

\item{sigma}{Numeric (positive); standard deviation of price simulated around current iteration.}

\item{alpha}{Numeric (\code{0 < alpha < 1}); how many (ratio of total) nearest observations to take for kernel regression.}
}
\value{
List with control parameters.
}
\description{
Control for Response Surface Algorithm
}
\seealso{
\code{\link{find_next_price}}
}

