% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_train_demand.R
\name{simulate_train_demand}
\alias{simulate_train_demand}
\title{Simulate train demand for whole train and whole selling period}
\usage{
simulate_train_demand(routes, model, beta, price, time_bp, nseats, keep_all)
}
\arguments{
\item{routes}{Output of \link{get_routes}. Do not build this argument manually.}

\item{model}{Model selection; either 1 or 2. See details.}

\item{beta}{Numeric matrix of model parameters. Each row contains parameters for one route.
The number of columns corresponds with the model selection. See details.}

\item{price}{Matrix of ticket prices. Each column represents one route, each row represents one time period.}

\item{time_bp}{Vector of times when the price changes. Values between 0 and 1.}

\item{keep_all}{Logical; Should the passengers that exceed the capacity be kept in the output?}

\item{nseat}{Number of seats in the train.}
}
\description{
Simulate the demand for whole train and whole selling period divided to multiple periods over which the price is constant.
}
\details{
If \code{keep_all} is TRUE all demanded tickets are returned no matter the capacity.
For each passenger the maximum acceptable price is returned.
}

